Se agrega Tipo de Pago y Tipo de Concepto

+++ src/main/java/py/com/fuentepy/appfinanzasBackend/model/PrestamoPagoModel.java
@@ -39,7 +39,7 @@ public class PrestamoPagoModel implements Serializable {
 
     @JsonProperty("tipoPagoId")
     @SerializedName("tipoPagoId")
-    private Integer tipoPagoId;
+    private Long tipoPagoId;
 
     @JsonProperty("tipoPagoNombre")
     @SerializedName("tipoPagoNombre")
